Move `/src` -> `/server`, remove views
