+++ server/app.js
@@ -4,16 +4,16 @@ const path = require("path")
 const cookieParser = require("cookie-parser")
 const logger = require("morgan")
 
-const indexRouter = require("../routes")
-const usersRouter = require("../routes/users")
+const indexRouter = require("./routes")
+const usersRouter = require("./routes/users")
 
 const root_dir = path.resolve(path.dirname("../"))
 
 const app = express()
 
 // view engine setup
-app.set("views", path.join(root_dir, "views"));
-app.set("view engine", "ejs");
+// app.set("views", path.join(root_dir, "views"));
+// app.set("view engine", "ejs");
 
 app.use(logger("dev"));
 app.use(express.json());
@@ -21,8 +21,8 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(root_dir, "public")));
 
-app.use("/", indexRouter);
-app.use("/users", usersRouter);
+// app.use("/", indexRouter);
+// app.use("/users", usersRouter);
 
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
