--- views/index.ejs
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta content="width=device-width,initial-scale=1" name="viewport">
-    <link rel="manifest" href="manifest.json">
-    <link href="/images/favicon.png" rel="icon" type="image/png">
-    <link rel="stylesheet" href="chota.min.css">
-    <title><%= title %></title>
-    <link rel="stylesheet" href="/styles/style.css"/>
-  </head>
-  <body>
-    <h1><%= title %></h1>
-    <p>Welcome to <%= title %></p>
-  </body>
-</html>
